pyproject.toml
[tool.poetry]
name = "note-vi-backend-fastapi"
version = "0.1.0"
description = ""
authors = ["Viva <sovestyaaaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.0"}
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
fastapi-users = "^12.1.3"
fastapi-users-db-sqlalchemy = "^6.0.1"
setuptools = "^69.1.1"
just-sh = "^0.0.2"
flower = "^2.0.1"
fastapi-cache2 = "^0.2.1"
celery = "^5.3.6"
redis = "^5.0.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
asynclog = "^0.1.7"
aiofiles = "^23.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode="auto"
filterwarnings=[
  "ignore::DeprecationWarning",
]
